OCR/requirement.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.14
chardet==3.0.4
charset-normalizer==3.4.2
click==8.2.1
colorlog==6.9.0
cssselect==1.3.0
cssutils==2.11.1
dataclasses-json==0.6.7
decorator==5.2.1
deep-translator==1.11.4
deepl==1.22.0
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
filelock==3.18.0
Flask==3.1.1
frozenlist==1.7.0
fsspec==2025.7.0
ftfy==6.3.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.2
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
GPUtil==1.4.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.9.0
h2==3.2.0
hf-xet==1.1.7
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
httpx-sse==0.4.1
huggingface-hub==0.34.3
hyperframe==5.2.0
idna==2.10
imageio==2.37.0
imagesize==1.4.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langsmith==0.4.8
lazy_loader==0.4
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
more-itertools==10.7.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
numpy==2.2.6
openai==1.99.9
opencv-contrib-python==4.10.0.84
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.11.2
packaging==25.0
paddleocr==3.1.0
paddlepaddle==3.1.0
paddlex==3.1.3
pandas==2.3.1
pillow==11.3.0
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
pypdfium2==4.30.0
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.11.0
tokenizers==0.21.4
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0
